+++ docs/CHANGELOG.md
@@ -11,3 +11,4 @@ eg h2GIS.getSpatialTable("h2gis").columns("land", "st_precisionreducer(st_transf
 - Filter null or empty table name before executing the drop table command 
 - Add getExtent method on datasource
 - Fix isEmpty when the table doesn't have any columns
+- Update H2 to 2.2.224 and fix PostGIS tests